job run every 5 mint to update database

+++ db/migrate/20231202111422_create_messages.rb
@@ -2,7 +2,7 @@ class CreateMessages < ActiveRecord::Migration[7.1]
   def change
     create_table :messages do |t|
       t.string :message_body
-      t.integer :message_number
+      t.decimal :message_number
       t.integer :chat_id
 
       t.timestamps
